feat(identity): enable project scope to take id or name
`from_env` now prefers id over name for both project and project domain.
To do this we had to change the signature of the setter functions for
the project scope.

The reason for this change is that our openstack server did not accept
just the project name and domain, but rather needed us to send the
project id and corresponding project domain id.

BREAKING_CHANGE: `with_project_scope` and `set_project_scope` now takes
`IdOrName` instead of string which is turned into a name.