feat: Add support for clouds-public.yaml and secure.yaml.
Before this change, only clouds.yaml was supported when getting a session
via `from_config`.

This change introduces support for clouds-public.yaml and secure.yaml.

It was not trivial to use serde_derive to parse the files individually
as all 3 files may contain all or none of the configuration, as long as
the merged result is valid configuration.
So the strategy here was to parse the 3 files into serde_yaml::Value
structs and merge them, and then convert them to the serde_derive
structs.